moving the catalog and includes structure into Corpus, and making the resultsDir a parameter
The Corpus constructor takes a parameter for the directory, and constructs in Includes and XMLCatalog as well as reads in documents.
The Includes constructor reads the files in a local includes directory, but must reconstruct the Titling (issue filed on this).
The XMLCatalog constructor reads in a local catalog file from disk.
The Includes and XMLCatalog classes now have their own write methods
Case Q added to test loading the results corpus from case G.